Split gateway into module-specific handlers
